Reimplement SC_FS as singleton; cleanup class
- Change to singleton with instance() accessor. I thought the use
of singleton would be justified here because it is a global object
with complex state. As a result some methods became non-static
(getDirectory most notably). gDirectoryMap became mDirectoryMap,
a private member
- Remove some apparently unneeded header includes in source file
- Rename "getXDirectory" methods to "defaultXDirectory" and make
them private instead of static
- Organize source file to reflect these changes
- Move setDirectory source into header
- Put SC_DirectoryName within SC_Filesystem namespace for clarity.
Rename to DirName. This results in some long references in certain
places, but with using-decls it's manageable. Still considering
the alternative of adding "Dir" to the end of each enum member
and eliminating the extra namespace.